renamed the client-edit-page to EditClientsPage
