+++ front-end-tests/cypress/pages/editRoomsPage.js
